充值管理

+++ czb-bn/bn-dao/src/main/java/com/cqut/czb/bn/dao/mapper/PetrolSalesRecordsMapperExtra.java
@@ -38,7 +38,7 @@ public interface PetrolSalesRecordsMapperExtra {
 
     int selectCountByWaitRecharge(@Param("list") List recordIds);
 
-    int getPetrolRechargeListTotalMoney(PetrolRechargeInputDTO inputDTO);
+    double getPetrolRechargeListTotalMoney(PetrolRechargeInputDTO inputDTO);
 
     int recharge(@Param("recordId") String recordId);
